+++ packages/use-wallet/src/index.ts
@@ -1,4 +1,5 @@
-export { WalletManager, WalletManagerConfig } from './manager'
+export { LogLevel } from './logger'
+export { WalletManager, WalletManagerConfig, WalletManagerOptions } from './manager'
 export { NetworkId } from './network'
 export { State, WalletState, defaultState } from './store'
 export { StorageAdapter } from './storage'