+++ packages/use-wallet/src/network.ts
@@ -4,9 +4,8 @@ export enum NetworkId {
   MAINNET = 'mainnet',
   TESTNET = 'testnet',
   BETANET = 'betanet',
-  LOCALNET = 'localnet',
-  VOITEST = 'voitest',
-  ARAMIDMAIN = 'aramidmain'
+  FNET = 'fnet',
+  LOCALNET = 'localnet'
 }
 
 export function isValidNetworkId(networkId: any): networkId is NetworkId {
@@ -45,11 +44,10 @@ export function isNetworkConfigMap(config: NetworkConfig): config is NetworkConf
 export type NetworkConfig = Partial<AlgodConfig> | Partial<Record<NetworkId, Partial<AlgodConfig>>>
 
 export const nodeServerMap = {
-  [NetworkId.MAINNET]: 'https://mainnet-api.algonode.cloud',
-  [NetworkId.TESTNET]: 'https://testnet-api.algonode.cloud',
-  [NetworkId.BETANET]: 'https://betanet-api.algonode.cloud',
-  [NetworkId.VOITEST]: 'https://testnet-api.voi.nodly.io',
-  [NetworkId.ARAMIDMAIN]: 'https://algod.aramidmain.a-wallet.net'
+  [NetworkId.MAINNET]: 'https://mainnet-api.4160.nodely.dev',
+  [NetworkId.TESTNET]: 'https://testnet-api.4160.nodely.dev',
+  [NetworkId.BETANET]: 'https://betanet-api.4160.nodely.dev',
+  [NetworkId.FNET]: 'https://fnet-api.4160.nodely.dev'
 }
 
 export function createDefaultNetworkConfig(): NetworkConfigMap {